+++ src/middlewares/createInsightsMiddleware.ts
@@ -1,6 +1,6 @@
 import { InsightsClient, InsightsClientMethod, Middleware } from '../types';
 import { getInsightsAnonymousUserTokenInternal } from '../helpers';
-import { warning, noop, getAppIdAndApiKey } from '../lib/utils';
+import { warning, noop, getAppIdAndApiKey, find } from '../lib/utils';
 
 export type InsightsEvent = {
   insightsMethod?: InsightsClientMethod;
@@ -38,26 +38,36 @@ export const createInsightsMiddleware: CreateInsightsMiddleware = props => {
     _insightsClient === null ? (noop as InsightsClient) : _insightsClient;
 
   return ({ instantSearchInstance }) => {
-    insightsClient('_get', '_hasCredentials', (hasCredentials: boolean) => {
-      if (!hasCredentials) {
-        const [appId, apiKey] = getAppIdAndApiKey(instantSearchInstance.client);
-        insightsClient('_get', '_userToken', (userToken: string) => {
-          warning(
-            !userToken,
-            `You set userToken before \`createInsightsMiddleware()\` and it is ignored.
-Please set the token after the \`createInsightsMiddleware()\` call.
+    const [appId, apiKey] = getAppIdAndApiKey(instantSearchInstance.client);
+    let queuedUserToken: string | undefined = undefined;
+    let userTokenBeforeInit: string | undefined = undefined;
 
-createInsightsMiddleware({ /* ... */ });
-
-insightsClient('setUserToken', 'your-user-token');
-// or
-aa('setUserToken', 'your-user-token');
-            `
-          );
-        });
-        insightsClient('init', { appId, apiKey });
-      }
+    if (Array.isArray(insightsClient.queue)) {
+      // Context: The umd build of search-insights is asynchronously loaded by the snippet.
+      //
+      // When user calls `aa('setUserToken', 'my-user-token')` before `search-insights` is loaded,
+      // ['setUserToken', 'my-user-token'] gets stored in `aa.queue`.
+      // Whenever `search-insights` is finally loaded, it will process the queue.
+      //
+      // But here's the reason why we handle it here:
+      // At this point, even though `search-insights` is not loaded yet,
+      // we still want to read the token from the queue.
+      // Otherwise, the first search call will be fired without the token.
+      [, queuedUserToken] =
+        find(
+          insightsClient.queue.slice().reverse(),
+          ([method]) => method === 'setUserToken'
+        ) || [];
+    }
+    insightsClient('_get', '_userToken', (userToken: string) => {
+      // If user has called `aa('setUserToken', 'my-user-token')` before creating
+      // the `insights` middleware, we store them temporarily and
+      // set it later on.
+      //
+      // Otherwise, the `init` call might override it with anonymous user token.
+      userTokenBeforeInit = userToken;
     });
+    insightsClient('init', { appId, apiKey });
 
     return {
       onStateChange() {},
@@ -76,28 +86,19 @@ aa('setUserToken', 'your-user-token');
           .getHelper()!
           .setQueryParameter('clickAnalytics', true);
 
-        if (hasInsightsClient) {
+        const anonymousUserToken = getInsightsAnonymousUserTokenInternal();
+        if (hasInsightsClient && anonymousUserToken) {
           // When `aa('init', { ... })` is called, it creates an anonymous user token in cookie.
           // We can set it as userToken.
-          setUserTokenToSearch(getInsightsAnonymousUserTokenInternal());
+          setUserTokenToSearch(anonymousUserToken);
         }
 
-        if (Array.isArray((insightsClient as any).queue)) {
-          // Context: The umd build of search-insights is asynchronously loaded by the snippet.
-          //
-          // When user calls `aa('setUserToken', 'my-user-token')` before `search-insights` is loaded,
-          // ['setUserToken', 'my-user-token'] gets stored in `aa.queue`.
-          // Whenever `search-insights` is finally loaded, it will process the queue.
-          //
-          // But here's the reason why we handle it here:
-          // At this point, even though `search-insights` is not loaded yet,
-          // we still want to read the token from the queue.
-          // Otherwise, the first search call will be fired without the token.
-          (insightsClient as any).queue.forEach(([method, firstArgument]) => {
-            if (method === 'setUserToken') {
-              setUserTokenToSearch(firstArgument);
-            }
-          });
+        // We consider the `userToken` coming from a `init` call to have a higher
+        // importance than the one coming from the queue.
+        if (userTokenBeforeInit) {
+          insightsClient('setUserToken', userTokenBeforeInit);
+        } else if (queuedUserToken) {
+          insightsClient('setUserToken', queuedUserToken);
         }
 
         // This updates userToken which is set explicitly by `aa('setUserToken', userToken)`